models/echomimic_v3_requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── PyTorch (install CUDA version separately — see README) ─────────────────
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# ── Diffusers / Transformers ───────────────────────────────────────────────
diffusers>=0.30.1
transformers>=4.46.2
accelerate>=0.25.0
safetensors>=0.4.0
huggingface_hub>=0.24.0
sentencepiece>=0.2.0

# ── EchoMimicV3 / Wan2.1 deps ─────────────────────────────────────────────
einops>=0.8.0
timm>=0.9.0
omegaconf>=2.3.0
tomesd>=0.1.3
torchsde>=0.2.6
# decord: no ARM64/macOS wheel — only used in infer_flash.py CLI, not by our server
# decord>=0.6.0
imageio>=2.28.0
imageio-ffmpeg>=0.4.9
scikit-image>=0.21.0
albumentations>=1.3.0
ftfy>=6.1.0
func_timeout>=4.3.5
mmgp>=0.1.0

# ── Audio ──────────────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.1
pyloudnorm>=0.1.1

# ── Video ──────────────────────────────────────────────────────────────────
moviepy==2.2.1
opencv-python>=4.8.0

# ── Face detection ─────────────────────────────────────────────────────────
retina-face==0.0.17
onnxruntime>=1.16.0

# ── Misc ───────────────────────────────────────────────────────────────────
Pillow>=10.0.0
numpy
tensorboard
datasets
beautifulsoup4

# ── Transitive deps missing from sub-package installs ─────────────────────
antlr4-python3-runtime==4.9.3
PyYAML>=6.0
regex>=2022.0
httpx>=0.27.0
tokenizers>=0.19.0
cffi>=1.0
lazy_loader>=0.4
scipy>=1.9.0
scikit-learn>=1.1.0
joblib>=1.0
numba>=0.51.0
audioread>=2.1.9
pooch>=1.1
soxr>=0.3.2
msgpack>=1.0
decorator>=4.3.0
starlette>=0.37.0
pydantic>=2.0
anyio>=4.0

# TensorFlow — Apple Silicon (ARM64) does not ship tensorflow==2.15.0.
# Use tensorflow-macos on Apple Silicon; standard tensorflow on Linux/CUDA.
# Detect at install time with the helper below, or install manually:
#   macOS ARM64 : pip install tensorflow-macos
#   Linux/CUDA  : pip install tensorflow==2.15.0
tensorflow-macos; sys_platform == "darwin" and platform_machine == "arm64"
tensorflow==2.15.0; sys_platform != "darwin"
